+++ NEWS.md
@@ -1,3 +1,7 @@
+# 2019-06
+
+* split_paths() now recognises server paths and keeps "//server" as one part
+
 # kwb.file 0.1.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.